Moved action types to their own file
